// src/api/generated/model/contestTeamAnalytics.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { ContestTeamMemberAnalytics } from './contestTeamMemberAnalytics';

export interface ContestTeamAnalytics {
  contest_team_id: string;
  name: string;
  score?: number;
  members?: ContestTeamMemberAnalytics[];
  total_submissions?: number;
  accepted_submission?: number;
  wrong_answer?: number;
  time_limit_exceeded?: number;
  runtime_error?: number;
  compilation_error?: number;
  memory_limit_exceeded?: number;
  system_error?: number;
  pending_submission?: number;
}